APIClientを修正（endpointを指定できるように）
