+++ SwiftUIQiitaClientMVVMTest/Utilities/APIClient.swift
@@ -12,16 +12,12 @@ import Combine
 protocol APIClient {
     var session: URLSession { get }
     var baseURL: String { get }
-    var path: String { get }
-    var method: String { get }
-    var headers: [String: String]? { get }
-    func body() throws -> Data?
 }
 
 extension APIClient {
-    func call<T>() -> AnyPublisher<T, APIError> where T: Decodable {
+    func call<T>(endPoint: APICall) -> AnyPublisher<T, APIError> where T: Decodable {
         do {
-            let request = try urlRequest(baseURL: baseURL)
+            let request = try endPoint.urlRequest(baseURL: baseURL)
             return session.dataTaskPublisher(for: request)
             .mapError { error in
                 .network(description: error.localizedDescription)
@@ -34,16 +30,5 @@ extension APIClient {
             let apiError = error as! APIError
             return Fail(error: apiError).eraseToAnyPublisher()
         }
-    }
-    
-    private func urlRequest(baseURL: String) throws -> URLRequest {
-        guard let url = URL(string: baseURL + path) else {
-            throw APIError.invalidURL(description: "APIError invalid URL!")
-        }
-        var request = URLRequest(url: url)
-        request.httpMethod = method
-        request.allHTTPHeaderFields = headers
-        request.httpBody = try body()
-        return request
     }
 }
