+++ SwiftUIQiitaClientMVVMTest/Utilities/HTTPMethod.swift
@@ -1,0 +1,16 @@
+//
+//  HTTPMethod.swift
+//  SwiftUIQiitaClientMVVMTest
+//
+//  Created by Shota Nishizawa on 2020/04/18.
+//  Copyright © 2020 Shota Nishizawa. All rights reserved.
+//
+
+import Foundation
+
+enum HTTPMethod: String {
+    case GET
+    case POST
+    case PUT
+    case DELETE
+}
